requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.5
jinja2==3.1.4
pybgcode-jneilliii==0.2.1
prusa-connect-sdk-client[cli]>=1.0.0
